requirements.txt
# requirements.txt
loguru
pyyaml
pandas
openpyxl
scrapy
markdownify
selenium
requests
